Program4/requirements.txt
Django
whitenoise
azure-storage-blob
azure-cosmos
django-dotenv

